fix(desktop): satisfy clipboard CI checks
