+++ desktop/src-tauri/src/commands/mod.rs
@@ -12,6 +12,7 @@ mod canvas;
 mod channel_templates;
 mod channel_window;
 mod channels;
+mod clipboard;
 mod dms;
 mod engrams;
 mod export_util;
@@ -61,6 +62,7 @@ pub use canvas::*;
 pub use channel_templates::*;
 pub use channel_window::*;
 pub use channels::*;
+pub use clipboard::*;
 pub use dms::*;
 pub use engrams::*;
 pub use global_agent_config::*;
